renamed namespace from PermitService to Permits
