+++ src/UKHO.S100PermitService.Common/IO/IPermitReaderWriter.cs
@@ -1,4 +1,4 @@
-﻿using UKHO.S100PermitService.Common.Models.PermitService;
+﻿using UKHO.S100PermitService.Common.Models.Permits;
 
 namespace UKHO.S100PermitService.Common.IO
 {
