+++ src/UKHO.S100PermitService.Common/Services/PermitService.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using UKHO.S100PermitService.Common.Events;
 using UKHO.S100PermitService.Common.IO;
-using UKHO.S100PermitService.Common.Models.PermitService;
+using UKHO.S100PermitService.Common.Models.Permits;
 
 namespace UKHO.S100PermitService.Common.Services
 {
