+++ tests/UKHO.S100PermitService.Common.UnitTests/IO/PermitReaderWriterTests.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using System.IO.Abstractions;
 using UKHO.S100PermitService.Common.IO;
-using UKHO.S100PermitService.Common.Models.PermitService;
+using UKHO.S100PermitService.Common.Models.Permits;
 
 namespace UKHO.S100PermitService.Common.UnitTests.IO
 {
